fix: Fixed member form input validation requirement to have either email or phone number
Fixes #104
